+++ src/infinitime/bluetooth/uuids.rs
@@ -8,6 +8,9 @@ pub const CHR_BATTERY_LEVEL: Uuid = uuid!("00002a19-0000-1000-8000-00805f9b34fb"
 pub const CHR_FIRMWARE_REVISION: Uuid = uuid!("00002a26-0000-1000-8000-00805f9b34fb");
 pub const CHR_HEART_RATE: Uuid = uuid!("00002a37-0000-1000-8000-00805f9b34fb");
 
+pub const CHR_NEW_ALERT: Uuid = uuid!("00002a46-0000-1000-8000-00805f9b34fb");
+pub const CHR_NOTIFICATION_EVENT: Uuid = uuid!("00020001-78fc-48fe-8e23-433b3a1942d0");
+
 pub const CHR_FWUPD_CONTROL_POINT: Uuid = uuid!("00001531-1212-efde-1523-785feabcd123");
 pub const CHR_FWUPD_PACKET: Uuid = uuid!("00001532-1212-efde-1523-785feabcd123");