+++ src/infinitime/mod.rs
@@ -1,7 +1,8 @@
 pub mod bluetooth;
 pub mod freedesktop;
 pub mod github;
+mod utils;
 
 pub use bluetooth as bt;
-pub use freedesktop as fdo; 
+pub use freedesktop as fdo;
 pub use github as gh;